app/server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
certifi==2023.5.7
    # via requests
charset-normalizer==2.1.1
    # via requests
django==4.2.3
    # via
    #   -r requirements.in
    #   djangorestframework
django-baton==2.7.1
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-prometheus==2.2.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
idna==3.4
    # via requests
prometheus-client==0.17.1
    # via django-prometheus
psycopg2-binary==2.9.1
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
requests==2.28.1
    # via -r requirements.in
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
urllib3==1.26.16
    # via requests
